+++ test/functional/workspace/test_workspace.py
@@ -47,48 +47,3 @@ def test_metabuild():
     assert "Adding project app1" in c.out
     c.run_command("cmake --build --preset conan-release")
     # it doesn't fail
-
-
-def test_relative_paths():
-    # This is using the meta-project
-    c = TestClient()
-    c.run("new cmake_lib -d name=liba --output=liba")
-    c.run("new cmake_exe -d name=app1 -d requires=liba/0.1 --output=app1")
-    c.run("new cmake_lib -d name=libb --output=other/libb")
-    c.run("new cmake_exe -d name=app2 -d requires=libb/0.1 --output=other/app2")
-    c.run("workspace init mywks")
-    c.run("workspace init otherwks")
-    # cd mywks
-    with c.chdir("mywks"):
-        c.run("workspace add ../liba")
-        c.run("workspace add ../app1 --product")
-        c.run("workspace info")
-        expected = textwrap.dedent("""\
-            products
-              ../app1
-            editables
-              app1/0.1
-                path: ../app1
-              liba/0.1
-                path: ../liba
-            """)
-        assert expected in c.out
-        c.run("graph info --requires=app1/0.1")
-        c.assert_listed_require({"app1/0.1": "Editable", "liba/0.1": "Editable"})
-    # cd otherwks
-    with c.chdir("otherwks"):
-        c.run("workspace add ../other/libb")
-        c.run("workspace add ../other/app2 --product")
-        c.run("workspace info")
-        expected = textwrap.dedent("""\
-            products
-              ../other/app2
-            editables
-              app2/0.1
-                path: ../other/app2
-              libb/0.1
-                path: ../other/libb
-            """)
-        assert expected in c.out
-        c.run("graph info --requires=app2/0.1")
-        c.assert_listed_require({"app2/0.1": "Editable", "libb/0.1": "Editable"})
